# internal/platform/httpserver/openapi.yaml
openapi: 3.0.3
info:
  title: Backend CRM Piposmart
  version: 0.15.4-sprint-14f-identity
  description: API CRM Piposmart dengan Partner Type, Partner (Encrypted & Masked Bank Account), PIC Assignment, Call Interactions, Lead Referrals, Subscription Order, Reconciliation Sprint 11, Partner Commission Sprint 12, penambahan route trash/unscoped owner-outlet dan rekap status langganan outlet Sprint 14b/14c, histori import persisten dengan viewer/download file asli Sprint 14 history, route /all dan /all-deleted lintas modul untuk kebutuhan frontend mengambil full dataset tanpa pagination, serta penambahan manajemen akun Admin dan Supervisor beserta reset password lintas role pada Sprint 14f.
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/subscription-orders:
    get:
      tags: [Subscription]
      summary: List subscription order
      operationId: listSubscriptionOrders
      description: Admin melihat seluruh order. Supervisor dan Sales hanya melihat order pada owner/lead dalam cakupan ownership mereka. Revenue top-up tetap dihitung dari wallet payment, sedangkan order dipakai untuk aktivasi subscription dan performa closing setelah reconciliation.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/SearchQuery"
        - name: owner_id
          in: query
          schema:
            type: integer
            format: int64
        - name: closing_id
          in: query
          schema:
            type: integer
            format: int64
        - name: sales_id
          in: query
          schema:
            type: integer
            format: int64
        - name: supervisor_id
          in: query
          schema:
            type: integer
            format: int64
        - name: plan_id
          in: query
          schema:
            type: integer
            format: int64
        - name: status
          in: query
          schema:
            type: string
            enum: [PAID, RECONCILED, REJECTED, CANCELED]
        - name: purchased_from
          in: query
          schema:
            type: string
            format: date
        - name: purchased_to
          in: query
          schema:
            type: string
            format: date
        - name: sort
          in: query
          schema:
            type: string
            example: -purchased_at
          description: Field yang didukung yaitu purchased_at, created_at, updated_at, final_amount, code. Gunakan prefix minus untuk descending.
        - $ref: "#/components/parameters/PageQuery"
        - $ref: "#/components/parameters/LimitQuery"
      responses:
        "200":
          description: List subscription order berhasil diambil
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
        "400":
          description: Query filter atau sort invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /api/v1/subscription-orders/{order_id}:
    get:
      tags: [Subscription]
      summary: Detail subscription order
      operationId: getSubscriptionOrder
      security:
        - bearerAuth: []
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Detail subscription order berhasil diambil
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
        "404":
          description: Order tidak ditemukan atau tidak dapat diakses user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /api/v1/owners/{owner_id}/subscription-orders:
    post:
      tags: [Subscription]
      summary: Membuat subscription order dari wallet owner
      operationId: createOwnerSubscriptionOrder
      description: Admin-only. Order mendebit wallet secara atomic, membuat subscription dan period, lalu otomatis membuat reconciliation jika closing_id dikirim. Wajib mengirim idempotency_key atau external_reference.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/OwnerID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubscriptionOrderRequest"
      responses:
        "201":
          description: Subscription order berhasil dibuat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
        "400":
          description: Payload invalid, promo tidak eligible, saldo tidak cukup, closing mismatch, atau idempotency kosong
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: User bukan Admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Owner, wallet, plan, promo, outlet, atau closing tidak ditemukan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /api/v1/subscription-orders/{order_id}/reconcile:
    post:
      tags: [Subscription]
      summary: Manual reconciliation subscription order dengan closing
      operationId: reconcileSubscriptionOrder
      description: Admin dan Supervisor dapat melakukan confirm/reject. Confirm mengaktifkan status reconciliation dan closing untuk KPI; reject tidak menghasilkan KPI/komisi.
      security:
        - bearerAuth: []
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ManualReconcileRequest"
      responses:
        "200":
          description: Reconciliation berhasil diproses
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
        "400":
          description: Action invalid atau closing tidak cocok dengan order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: User tidak memiliki akses reconciliation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: Order sudah direkonsiliasi
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /api/v1/subscriptions:
    get:
      tags: [Subscription]
      summary: List subscription aktif/expired owner
      operationId: listSubscriptions
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/SearchQuery"
        - name: owner_id
          in: query
          schema:
            type: integer
            format: int64
        - name: order_id
          in: query
          schema:
            type: integer
            format: int64
        - name: plan_id
          in: query
          schema:
            type: integer
            format: int64
        - name: status
          in: query
          schema:
            type: string
            enum: [ACTIVE, EXPIRED, CANCELED]
        - name: active_from
          in: query
          schema:
            type: string
            format: date
        - name: active_to
          in: query
          schema:
            type: string
            format: date
        - name: sort
          in: query
          schema:
            type: string
            example: -active_from
          description: Field yang didukung yaitu active_from, active_until, created_at, updated_at, code.
        - $ref: "#/components/parameters/PageQuery"
        - $ref: "#/components/parameters/LimitQuery"
      responses:
        "200":
          description: List subscription berhasil diambil
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
  /api/v1/subscriptions/{subscription_id}:
    get:
      tags: [Subscription]
      summary: Detail subscription
      operationId: getSubscription
      security:
        - bearerAuth: []
      parameters:
        - name: subscription_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Detail subscription berhasil diambil
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
        "404":
          description: Subscription tidak ditemukan atau tidak dapat diakses user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /api/v1/reconciliations:
    get:
      tags: [Subscription]
      summary: List reconciliation order dan closing
      operationId: listSubscriptionReconciliations
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/SearchQuery"
        - name: order_id
          in: query
          schema:
            type: integer
            format: int64
        - name: closing_id
          in: query
          schema:
            type: integer
            format: int64
        - name: owner_id
          in: query
          schema:
            type: integer
            format: int64
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING, CONFIRMED, REJECTED]
        - name: sort
          in: query
          schema:
            type: string
            example: -created_at
          description: Field yang didukung yaitu created_at, updated_at, confirmed_at, amount_difference, code.
        - $ref: "#/components/parameters/PageQuery"
        - $ref: "#/components/parameters/LimitQuery"
      responses:
        "200":
          description: List reconciliation berhasil diambil
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
  /api/v1/reconciliation-issues:
    get:
      tags: [Subscription]
      summary: List issue reconciliation yang perlu review
      operationId: listReconciliationIssues
      description: Digunakan untuk menemukan hanging transaction, order tanpa closing, atau mismatch reconciliation.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/SearchQuery"
        - name: order_id
          in: query
          schema:
            type: integer
            format: int64
        - name: owner_id
          in: query
          schema:
            type: integer
            format: int64
        - name: issue_type
          in: query
          schema:
            type: string
            enum: [HANGING_ORDER, CLOSING_MISMATCH, MANUAL_REVIEW]
        - name: status
          in: query
          schema:
            type: string
            enum: [OPEN, RESOLVED]
        - name: sort
          in: query
          schema:
            type: string
            example: -detected_at
          description: Field yang didukung yaitu detected_at, created_at, updated_at, code.
        - $ref: "#/components/parameters/PageQuery"
        - $ref: "#/components/parameters/LimitQuery"
      responses:
        "200":
          description: List issue reconciliation berhasil diambil
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
...
  /api/v1/subscription-orders/all:
    get:
      tags: [Subscription]
      summary: List semua subscription order tanpa pagination
      operationId: listSubscriptionOrdersAll
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/SearchQuery"
        - name: owner_id
          in: query
          schema:
            type: integer
            format: int64
        - name: outlet_id
          in: query
          schema:
            type: integer
            format: int64
        - name: closing_id
          in: query
          schema:
            type: integer
            format: int64
        - name: sales_id
          in: query
          schema:
            type: integer
            format: int64
        - name: supervisor_id
          in: query
          schema:
            type: integer
            format: int64
        - name: plan_id
          in: query
          schema:
            type: integer
            format: int64
        - name: status
          in: query
          schema:
            type: string
        - name: purchased_from
          in: query
          schema:
            type: string
            format: date
        - name: purchased_to
          in: query
          schema:
            type: string
            format: date
        - name: sort
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List semua subscription order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
  /api/v1/subscription-orders/all-deleted:
    get:
      tags: [Subscription]
      summary: Alias full subscription order untuk kompatibilitas frontend
      operationId: listSubscriptionOrdersAllDeleted
      description: Modul subscription order list saat ini tidak memakai soft delete terpisah pada route ini. Route ini adalah alias kompatibilitas frontend.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/SearchQuery"
        - name: owner_id
          in: query
          schema:
            type: integer
            format: int64
        - name: outlet_id
          in: query
          schema:
            type: integer
            format: int64
        - name: closing_id
          in: query
          schema:
            type: integer
            format: int64
        - name: sales_id
          in: query
          schema:
            type: integer
            format: int64
        - name: supervisor_id
          in: query
          schema:
            type: integer
            format: int64
        - name: plan_id
          in: query
          schema:
            type: integer
            format: int64
        - name: status
          in: query
          schema:
            type: string
        - name: purchased_from
          in: query
          schema:
            type: string
            format: date
        - name: purchased_to
          in: query
          schema:
            type: string
            format: date
        - name: sort
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List semua subscription order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
  /api/v1/subscriptions/all:
    get:
      tags: [Subscription]
      summary: List semua subscription tanpa pagination
      operationId: listSubscriptionsAll
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/SearchQuery"
        - name: owner_id
          in: query
          schema:
            type: integer
            format: int64
        - name: order_id
          in: query
          schema:
            type: integer
            format: int64
        - name: plan_id
          in: query
          schema:
            type: integer
            format: int64
        - name: status
          in: query
          schema:
            type: string
            enum: [ACTIVE, EXPIRED, CANCELED]
        - name: active_from
          in: query
          schema:
            type: string
            format: date
        - name: active_to
          in: query
          schema:
            type: string
            format: date
        - name: sort
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List semua subscription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
  /api/v1/subscriptions/all-deleted:
    get:
      tags: [Subscription]
      summary: Alias full subscription untuk kompatibilitas frontend
      operationId: listSubscriptionsAllDeleted
      description: Modul subscription list saat ini tidak memakai soft delete terpisah pada route ini. Route ini adalah alias kompatibilitas frontend.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/SearchQuery"
        - name: owner_id
          in: query
          schema:
            type: integer
            format: int64
        - name: order_id
          in: query
          schema:
            type: integer
            format: int64
        - name: plan_id
          in: query
          schema:
            type: integer
            format: int64
        - name: status
          in: query
          schema:
            type: string
            enum: [ACTIVE, EXPIRED, CANCELED]
        - name: active_from
          in: query
          schema:
            type: string
            format: date
        - name: active_to
          in: query
          schema:
            type: string
            format: date
        - name: sort
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List semua subscription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
  /api/v1/reconciliations/all:
    get:
      tags: [Subscription]
      summary: List semua reconciliation tanpa pagination
      operationId: listSubscriptionReconciliationsAll
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/SearchQuery"
        - name: order_id
          in: query
          schema:
            type: integer
            format: int64
        - name: closing_id
          in: query
          schema:
            type: integer
            format: int64
        - name: owner_id
          in: query
          schema:
            type: integer
            format: int64
        - name: sales_id
          in: query
          schema:
            type: integer
            format: int64
        - name: supervisor_id
          in: query
          schema:
            type: integer
            format: int64
        - name: status
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List semua reconciliation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
  /api/v1/reconciliations/all-deleted:
    get:
      tags: [Subscription]
      summary: Alias full reconciliation untuk kompatibilitas frontend
      operationId: listSubscriptionReconciliationsAllDeleted
      description: Modul reconciliation list saat ini tidak memakai soft delete terpisah pada route ini. Route ini adalah alias kompatibilitas frontend.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/SearchQuery"
        - name: order_id
          in: query
          schema:
            type: integer
            format: int64
        - name: closing_id
          in: query
          schema:
            type: integer
            format: int64
        - name: owner_id
          in: query
          schema:
            type: integer
            format: int64
        - name: sales_id
          in: query
          schema:
            type: integer
            format: int64
        - name: supervisor_id
          in: query
          schema:
            type: integer
            format: int64
        - name: status
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List semua reconciliation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
  /api/v1/reconciliation-issues/all:
    get:
      tags: [Subscription]
      summary: List semua reconciliation issue tanpa pagination
      operationId: listReconciliationIssuesAll
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/SearchQuery"
        - name: issue_type
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: owner_id
          in: query
          schema:
            type: integer
            format: int64
        - name: order_id
          in: query
          schema:
            type: integer
            format: int64
        - name: closing_id
          in: query
          schema:
            type: integer
            format: int64
        - name: sales_id
          in: query
          schema:
            type: integer
            format: int64
        - name: supervisor_id
          in: query
          schema:
            type: integer
            format: int64
        - name: sort
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List semua reconciliation issue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
  /api/v1/reconciliation-issues/all-deleted:
    get:
      tags: [Subscription]
      summary: Alias full reconciliation issue untuk kompatibilitas frontend
      operationId: listReconciliationIssuesAllDeleted
      description: Modul reconciliation issue list saat ini tidak memakai soft delete terpisah pada route ini. Route ini adalah alias kompatibilitas frontend.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/SearchQuery"
        - name: issue_type
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: owner_id
          in: query
          schema:
            type: integer
            format: int64
        - name: order_id
          in: query
          schema:
            type: integer
            format: int64
        - name: closing_id
          in: query
          schema:
            type: integer
            format: int64
        - name: sales_id
          in: query
          schema:
            type: integer
            format: int64
        - name: supervisor_id
          in: query
          schema:
            type: integer
            format: int64
        - name: sort
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List semua reconciliation issue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    OwnerID:
      name: owner_id
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
    SearchQuery:
      name: q
      in: query
      required: false
      schema:
        type: string
      description: Search global untuk kode, nama, telepon, brand, kota, dan provinsi.
...
    PageQuery:
      name: page
      in: query
      required: false
      schema:
        type: integer
        default: 1
        minimum: 1
    LimitQuery:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        default: 10
        minimum: 1
        maximum: 100
  schemas:
...
    ManualReconcileRequest:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [CONFIRM, REJECT]
          example: CONFIRM
        closing_id:
          type: integer
          format: int64
          nullable: true
          description: Wajib untuk action CONFIRM jika order belum punya closing.
          example: 9
        note:
          type: string
          example: Cocok dengan closing Sales Juli.
        reason:
          type: string
          example: Nominal atau owner tidak cocok.
...
    SuccessEnvelope:
      type: object
      required: [data, meta]
      properties:
        data:
          type: object
          additionalProperties: true
        meta:
          type: object
          required: [request_id]
          properties:
            request_id:
              type: string
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              nullable: true
            request_id:
              type: string
...